feat: tom-select fields (WIP)
Basic search is working, but plumbing to the backend is not there yet. Does not yet support all the features we use with select2. Next step is to flesh that out.

Running into some styling issues. The tom-select.bootstrap5 style breaks parcel's attempt at packaging due to a conflict of some sort with the $select-padding-y variable it sets. Works with default, but the select box widget does not style correctly and the pull-down to list options is entirely non-functional.